# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sourdough API
  description: |
    Sourdough is a starter application framework for AI to develop other apps, providing 
    enterprise-grade user management, multi-provider notification system, and 
    multi-LLM orchestration with council mode.
  version: 0.1.0
  contact:
    name: Sourdough Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      security:
        - sanctum: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /email-templates:
    get:
      tags: [Email Templates]
      summary: List email templates
      security:
        - sanctum: []
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailTemplate'
  /email-templates/{key}:
    get:
      tags: [Email Templates]
      summary: Get email template
      security:
        - sanctum: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template details
    put:
      tags: [Email Templates]
      summary: Update email template
      security:
        - sanctum: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                body_html:
                  type: string
                body_text:
                  type: string
      responses:
        '200':
          description: Template updated
...
  /notifications/{notification}:
    delete:
      tags: [Notifications]
      summary: Delete notification
      security:
        - sanctum: []
      parameters:
        - name: notification
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification deleted

  # ==================== Admin Notification Channels ====================
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      description: Laravel Sanctum session-based authentication
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        message:
          type: string
        data:
          type: object
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
...